crypto_backend/requirements.txt
fastapi
uvicorn
python-dotenv
alpaca-py
requests
